+++ src/theme/DocItem/Layout/index.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react" />
+import type { Props } from '@theme/DocItem/Layout';
+export default function DocItemLayout({ children }: Props): JSX.Element;
