+++ src/theme/DocItem/Layout/styles.module.css
@@ -7,15 +7,4 @@
   .docItemCol {
     max-width: 75% !important;
   }
-
-  /* Prevent hydration FOUC, as the mobile TOC needs to be server-rendered */
-  .tocMobile {
-    display: none;
-  }
-}
-
-@media print {
-  .tocMobile {
-    display: none;
-  }
 }
